[SYSTEMDS-3463] Add unique() built-in function
This patch converts the existing unique() function from a script to a built-in. The script-based approach is based on sorting, which is very expensive computationally, especially for large multiblock inputs. The new approach, on the other hand, is based on a new data sketch for the unique() function. This first patch creates the framework for the new unique sketch and implements the CP RowCol case; other cases - CP Row/Col and Spark RowCol/Row/Col - will be implemented in subsequent patches.

Tests:
    [X] Unit tests
    [ ] Integration tests
    [ ] N/A